refactor: re-consolidate model catalog modules and engine metadata URLs
