+++ basilisk/model_catalog/__init__.py
@@ -1,0 +1,1 @@
+"""Catalog-driven sampling policy and model detail text for catalog-fed models."""
